Take over from Codex
Fix up the buttons to a state worth checking in and merging over to my issue branch
