+++ src/reefcraft/ui/section.py
@@ -1,0 +1,17 @@
+# -----------------------------------------------------------------------------
+# Copyright (c) 2025 The Reefcraft Project.
+#
+# Licensed under the MIT License. See the LICENSE file for details.
+# -----------------------------------------------------------------------------
+
+"""Defines the main GUI layout as a side panel."""
+
+from reefcraft.utils.logger import logger
+
+
+class Section:
+    """A sub-section of the UI to show/hide."""
+
+    def __init__(self) -> None:
+        """Initialize the section TBD."""
+        logger.debug("Section initialized - not yet implemented")
