+++ src/reefcraft/ui/slider.py
@@ -6,39 +6,12 @@
 
 """Defines the main GUI layout as a side panel."""
 
+from collections.abc import Callable
+
 import pygfx as gfx
 
-from reefcraft.sim.state import SimState
-from reefcraft.utils.logger import logger
-
-
-class Widget:
-    """A base cass for all widget types."""
-
-    def __init__(self, top: int, left: int, width: int, height: int) -> None:
-        """Create the slider and add it to the given ``panel`` scene."""
-        self.top = top
-        self.left = left
-        self.width = width
-        self.height = height
-
-
-class ListLayout:
-    """A simple retained-mode slider widget."""
-
-    def __init__(self) -> None:
-        """Create the slider and add it to the given ``panel`` scene."""
-        self.widgets: list[Widget] = []
-        self.line_spacing = 10
-
-    def add_widget(self, widget: Widget) -> None:
-        """Append another widget to the list."""
-        self.widgets.append(widget)
-
-    @property
-    def height(self) -> int:
-        """Sum the heights of all the widgets plus the line spacing."""
-        return 100
+from reefcraft.ui.panel import Panel
+from reefcraft.ui.widget import Widget
 
 
 class Slider(Widget):
@@ -46,7 +19,7 @@ class Slider(Widget):
 
     def __init__(
         self,
-        panel: "Panel",
+        panel: Panel,
         *,
         left: int,
         top: int,
@@ -55,9 +28,7 @@ class Slider(Widget):
         min_value: float = 0.0,
         max_value: float = 1.0,
         value: float | None = None,
-        background_color: str = "#1E182B",
-        foreground_color: str = "#4343EC",
-        text_color: str = "#D8D8D8",
+        on_change: Callable[[float], None] | None = None,
     ) -> None:
         """Create the slider and add it to the given ``panel`` scene."""
         super().__init__(top, left, width, height)
@@ -65,16 +36,14 @@ class Slider(Widget):
         self.min = min_value
         self.max = max_value
         self.value = value if value is not None else min_value
-        self.background_color = background_color
-        self.foreground_color = foreground_color
-        self.text_color = text_color
+        self._on_change_callback = on_change
 
         self._dragging = False
 
         # Background mesh
         self._bg_mesh = gfx.Mesh(
             gfx.plane_geometry(width=width, height=height),
-            gfx.MeshBasicMaterial(color=background_color),
+            gfx.MeshBasicMaterial(color=self.theme.color),
         )
         if self._bg_mesh.material is not None:
             self._bg_mesh.material.pick_write = True
@@ -82,11 +51,11 @@ class Slider(Widget):
         # Foreground mesh showing the filled portion
         self._fg_mesh = gfx.Mesh(
             gfx.plane_geometry(width=1, height=height),
-            gfx.MeshBasicMaterial(color=foreground_color),
+            gfx.MeshBasicMaterial(color=self.theme.highlight_color),
         )
 
         # Text overlay (transparent background)
-        text_mat = gfx.TextMaterial(color=text_color)
+        text_mat = gfx.TextMaterial(color=self.theme.text_color)
         self._text = gfx.Text(str(self.value), text_mat)
 
         self.panel.scene.add(self._bg_mesh)
@@ -98,6 +67,14 @@ class Slider(Widget):
         self._bg_mesh.add_event_handler(self._on_mouse_up, "pointer_up")
 
         self._update_visuals()
+
+    def set_value(self, value: float) -> None:
+        """Changes the value and calls the callback to the data source."""
+        if value != self.value:
+            self.value = value
+            if self._on_change_callback:
+                self._on_change_callback(self.value)
+            self._update_visuals()
 
     # ------------------------------------------------------------------
     # Utility helpers
@@ -117,8 +94,7 @@ class Slider(Widget):
     def _set_from_screen_x(self, x: float) -> None:
         t = (x - self.left) / self.width
         t = max(0.0, min(1.0, t))
-        self.value = self.min + t * (self.max - self.min)
-        self._update_visuals()
+        self.set_value(self.min + t * (self.max - self.min))
 
     def _update_visuals(self) -> None:
         filled = max(0.0, min(1.0, self._percent))
@@ -154,51 +130,3 @@ class Slider(Widget):
             self._dragging = False
             event.target.release_pointer_capture(event.pointer_id)
             self._set_from_screen_x(event.x)
-
-
-class Section:
-    """A sub-section of the UI to show/hide."""
-
-    def __init__(self) -> None:
-        """Initialize the section TBD."""
-        logger.debug("Section initialized - not yet implemented")
-
-
-class Panel:
-    """A left-docked panel: covers left 300px of canvas height."""
-
-    def __init__(self, renderer: gfx.WgpuRenderer, width: int = 300, height: int = 1080) -> None:
-        """Initialize the panel and its correstponding 3D scene and ortho cameras."""
-        self.renderer = renderer
-        self.viewport = gfx.Viewport(renderer)
-
-        geom = gfx.plane_geometry(width=width, height=height, width_segments=1, height_segments=1)
-        mat = gfx.MeshBasicMaterial(color="#08080A")
-        mesh = gfx.Mesh(geom, mat)
-        mesh.local.position = (-((1920 / 2) - (300 / 2)), 0, -100)
-
-        # Block the picking for the background of the panel
-        if mesh.material is not None:
-            mesh.material.pick_write = True
-        mesh.add_event_handler(self._on_mouse_down, "pointer_down")
-        mesh.add_event_handler(self._on_mouse_up, "pointer_up")
-
-        self.scene = gfx.Scene()
-        self.camera = gfx.OrthographicCamera(width=1920, height=1080)
-        self.scene.add(mesh)
-
-        Slider(self, left=20, top=20, width=250, height=20, min_value=0, max_value=100, value=10)
-        Slider(self, left=20, top=50, width=250, height=20, min_value=0, max_value=100, value=70)
-        Slider(self, left=20, top=80, width=250, height=20, min_value=0, max_value=100, value=15)
-
-    def _on_mouse_down(self, event: gfx.PointerEvent) -> None:
-        """When the mouse is clicked in background of the panel, capture the mouse and block others."""
-        event.target.set_pointer_capture(event.pointer_id, self.renderer)
-
-    def _on_mouse_up(self, event: gfx.PointerEvent) -> None:
-        """Release the mouse on mouse up."""
-        event.target.release_pointer_capture(event.pointer_id)
-
-    def draw(self, state: SimState) -> None:
-        """Draw a solid rectangle on the left side of the UI scene."""
-        self.viewport.render(self.scene, self.camera)  # , flush=False)
