+++ tests/python/test_button.py
@@ -3,7 +3,7 @@ from pathlib import Path
 
 sys.path.append(str(Path(__file__).resolve().parents[2] / "src"))
 
-from reefcraft.views.button import Button, ButtonState
+from reefcraft.ui.button import Button, ButtonState
 
 
 class DummyTarget:
